+++ js/main.js
@@ -1,2 +1,4 @@
-import {allAdvertisements} from './main-function.js';
+import {ADVERTISEMENT_NUMBER, createAdvertisement} from './data.js';
+
+const allAdvertisements = new Array(ADVERTISEMENT_NUMBER).fill().map(() => createAdvertisement());
 // console.log(allAdvertisements);
